% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustCompositional-package.R
\docType{package}
\name{mclustCompositional-package}
\alias{mclustCompositional}
\alias{mclustCompositional-package}
\title{\strong{mclustCompositional} package}
\description{
An R package implementing model-based clustering and density
estimation for compositional data using transformation-based
Gaussian mixtures.
}
\author{
\strong{Maintainer}: Luca Scrucca \email{luca.scrucca@unibo.it} (\href{https://orcid.org/0000-0003-3826-0484}{ORCID})

Other contributors:
\itemize{
  \item Fedele Greco \email{fedele.greco@unibo.it} [contributor]
}

}
\keyword{internal}
